+++ usermgmt/src/cli_ssh_credential.rs
@@ -1,7 +1,9 @@
 use log::info;
 use once_cell::sync::OnceCell;
+use usermgmt_lib::cli::OptFilePath;
 use usermgmt_lib::prelude::*;
 
+use usermgmt_lib::ssh::SshKeyPair;
 use usermgmt_lib::{
     config::MgmtConfig,
     prelude::{anyhow, AppResult},
@@ -15,14 +17,21 @@ pub struct CliSshCredential {
     default_ssh_user: String,
     username: OnceCell<String>,
     password: OnceCell<String>,
+    ssh_key_path: Option<SshKeyPair>,
 }
 
 impl CliSshCredential {
-    pub fn new(config: &MgmtConfig) -> Self {
+    pub fn new(config: &MgmtConfig, on_which_sys: &OptFilePath) -> Self {
+        let ssh_key_path = on_which_sys
+            .as_ref()
+            .cloned()
+            .or_else(|| config.ssh_key_path.clone())
+            .map(SshKeyPair::from_one_path);
         Self {
             username: Default::default(),
             password: Default::default(),
             default_ssh_user: config.default_ssh_user.clone(),
+            ssh_key_path,
         }
     }
 }
@@ -41,8 +50,8 @@ impl SshCredentials for CliSshCredential {
     }
     fn password(&self) -> AppResult<&str> {
         let password = self.password.get_or_try_init(|| {
-            let maybe_password = user_input::cli_ask_for_password("Enter your SSH password: ")?;
-            maybe_password.ok_or_else(|| anyhow!("No password provided"))
+            let from_prompt = user_input::cli_ask_for_password("Enter your SSH password: ")?;
+            Ok::<String, AppError>(from_prompt.unwrap_or_default())
         })?;
 
         Ok(password)
@@ -73,5 +82,9 @@ impl SshCredentials for CliSshCredential {
             info!("{}. ssh key is chosen", user_choice);
             Ok(last_index)
         }
+    }
+
+    fn ssh_paths_pair_key(&self) -> Option<&SshKeyPair> {
+        self.ssh_key_path.as_ref()
     }
 }